test: set auth config
