+++ src/tests/constants/auth-tests.constants.ts
@@ -1,0 +1,4 @@
+export const EXTERNAL_ALTERNATIVE_ORIGINS = ['other.com', 'another.com'];
+export const EXTERNAL_ALTERNATIVE_ORIGINS_URLS = EXTERNAL_ALTERNATIVE_ORIGINS.map(
+	(url) => `https://${url}`
+);
